+++ framework/Config/Config.php
@@ -2,14 +2,29 @@
 
 namespace Framework\Config;
 
+use Framework\Facades\Env;
 use Framework\Facades\Path;
+use Framework\Test\SupportsTestModeInterface;
 
-class Config
+class Config implements SupportsTestModeInterface
 {
     private static ?array $env = null;
 
+    // Test mode
+    private static bool $isTestMode = false;
+    private static array $testValues = [];
+
     public static function env(string $name, mixed $default = null): mixed
     {
+        // Test mode
+        if (self::$isTestMode) {
+            if (!array_key_exists($name, self::$testValues)) {
+                return $default;
+            }
+
+            return self::$testValues[$name];
+        }
+
         // Check if the requested value is set as system environment variable
         $sysEnv = getenv($name);
         if ($sysEnv !== false) {
@@ -32,5 +47,17 @@ class Config
             return;
         }
         self::$env = ConfigReader::readFile(Path::join(__DIR__, '..', '..', '.env'));
+    }
+
+    // -------------------- Test mode -------------------- 
+
+    public static function useTestMode(): void
+    {
+        self::$isTestMode = Env::isTestRun();
+    }
+
+    public static function setTestValues(string $key, mixed $value): void
+    {
+        self::$testValues[$key] = $value;
     }
 }