+++ framework/Facades/Env.php
@@ -6,6 +6,11 @@ use Framework\Config\Config;
 
 class Env
 {
+    public static function isTestRun(): bool
+    {
+        return key_exists('isTestRun', $GLOBALS);
+    }
+
     public static function isCLI(): bool
     {
         return php_sapi_name() === 'cli';