+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 
 ### Added
 - Added class `CreateTableBlueprint` to create statements for generating tables in MySQL/MariaDB and SQLite
+- Added test mode to class `Config` so env values can be changed in test runs
 
 ## v2.5.0 - 17.03.2024 - Added statistic for amount and price development